+++ docs/CHANGELOG.rst
@@ -6,6 +6,24 @@ All notable changes to the Vector SIL Kit project shall be documented in this fi
 
 The format is based on `Keep a Changelog (http://keepachangelog.com/en/1.0.0/) <http://keepachangelog.com/en/1.0.0/>`_.
 
+
+[4.0.36] - Unreleased
+---------------------
+
+Fixed
+~~~~~
+
+- Fixed misbehavior of the sil-kit-system-controller in interactive mode on user input:
+  - The sil-kit-system-controller now triggers a Stop() in SystemState::Running or SystemState::Paused.
+  - The sil-kit-system-controller only triggers AbortSimulation when not SystemState::Running, SystemState::Paused, SystemState::Shutdown or SystemState::Aborting.
+
+Changed
+~~~~~~~
+
+- The final state handling of the sil-kit-system-controller in interactive mode on user input has changed:
+  - Old: The sil-kit-system-controller triggered AbortSimulation if the finalState was not received after 5s. 
+  - New: The sil-kit-system-controller retries receiving the finalState 3x5s. If this fails, the sil-kit-system-controller triggers AbortSimulation (if not already happened) and tries receiving the finalState 3x5s again. If this fails, the sil-kit-system-controller just terminates.
+
 [4.0.35] - 2023-09-04
 ---------------------